Rename the drawable caret. One caret position should match exactly one drawable caret.
