+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Beatmaps/Lyrics/Components/Lyrics/Carets/DrawableTimeTagCaret.cs
@@ -3,6 +3,7 @@
 
 #nullable disable
 
+using System;
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Game.Graphics;
@@ -13,7 +14,7 @@ using osuTK;
 
 namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Components.Lyrics.Carets;
 
-public partial class DrawableTimeTagEditCaret : DrawableCaret<TimeTagIndexCaretPosition>
+public partial class DrawableTimeTagCaret : DrawableCaret<TimeTagCaretPosition>
 {
     private const float triangle_width = 8;
 
@@ -25,7 +26,7 @@ public partial class DrawableTimeTagEditCaret : DrawableCaret<TimeTagIndexCaretP
 
     private readonly DrawableTextIndex drawableTextIndex;
 
-    public DrawableTimeTagEditCaret(DrawableCaretType type)
+    public DrawableTimeTagCaret(DrawableCaretType type)
         : base(type)
     {
         AutoSizeAxes = Axes.Both;
@@ -38,14 +39,23 @@ public partial class DrawableTimeTagEditCaret : DrawableCaret<TimeTagIndexCaretP
         };
     }
 
-    protected override void Apply(TimeTagIndexCaretPosition caret)
+    protected override void Apply(TimeTagCaretPosition caret)
     {
-        var textIndex = caret.Index;
-        Position = karaokeSpriteText.GetTextIndexPosition(textIndex);
+        var timeTag = caret.TimeTag;
+        var textIndex = timeTag.Index;
+        this.MoveTo(karaokeSpriteText.GetTimeTagPosition(timeTag), getMoveToDuration(Type), Easing.OutCubic);
         Origin = TextIndexUtils.GetValueByState(textIndex, Anchor.BottomLeft, Anchor.BottomRight);
 
         drawableTextIndex.State = textIndex.State;
-        drawableTextIndex.Colour = colours.GetEditTimeTagCaretColour();
+        drawableTextIndex.Colour = colours.GetRecordingTimeTagCaretColour(timeTag);
+
+        static double getMoveToDuration(DrawableCaretType type) =>
+            type switch
+            {
+                DrawableCaretType.Caret => 100,
+                DrawableCaretType.HoverCaret => 0,
+                _ => throw new ArgumentOutOfRangeException(nameof(type), type, null)
+            };
     }
 
     public override void TriggerDisallowEditEffect(LyricEditorMode editorMode)
